+++ core/dplugins/dimg/jpeg/dimgjpegplugin.cpp
@@ -154,7 +154,7 @@ int DImgJPEGPlugin::canRead(const QFileInfo& fileInfo, bool magic) const
         return 0;
     }
 
-    const int headerLen = 9;
+    const qint64 headerLen = 9;
 
     QByteArray header(headerLen, '\0');
