+++ core/dplugins/dimg/pgf/dimgpgfplugin.cpp
@@ -141,7 +141,7 @@ int DImgPGFPlugin::canRead(const QFileInfo& fileInfo, bool magic) const
         return 0;
     }
 
-    const int headerLen = 9;
+    const qint64 headerLen = 9;
 
     QByteArray header(headerLen, '\0');
