+++ core/dplugins/dimg/png/dimgpngplugin.cpp
@@ -139,7 +139,7 @@ int DImgPNGPlugin::canRead(const QFileInfo& fileInfo, bool magic) const
         return 0;
     }
 
-    const int headerLen = 9;
+    const qint64 headerLen = 9;
 
     QByteArray header(headerLen, '\0');
