+++ core/dplugins/dimg/tiff/dimgtiffplugin.cpp
@@ -148,7 +148,7 @@ int DImgTIFFPlugin::canRead(const QFileInfo& fileInfo, bool magic) const
         return 0;
     }
 
-    const int headerLen = 9;
+    const qint64 headerLen = 9;
 
     QByteArray header(headerLen, '\0');
